ATA-5200: management-console: reorganization of "mid" packages
