+++ prism-backend/management-console/src/main/scala/io/iohk/atala/prism/management/console/repositories/CredentialIssuancesRepository.scala
@@ -16,13 +16,13 @@ import io.iohk.atala.prism.management.console.repositories.daos.CredentialIssuan
 import io.iohk.atala.prism.management.console.repositories.daos._
 import io.iohk.atala.prism.utils.syntax.DBConnectionOps
 import io.scalaland.chimney.dsl._
-import io.iohk.atala.prism.metrics.{TimeMeasureMetric, TimeMeasureUtil}
-import io.iohk.atala.prism.metrics.TimeMeasureUtil.MeasureOps
+import io.iohk.atala.prism.metrics.TimeMeasureMetric
 import io.iohk.atala.prism.management.console.models.CredentialTypeWithRequiredFields
 import io.iohk.atala.prism.management.console.repositories.CredentialIssuancesRepository.{
   CreateCredentialBulk,
   CreateCredentialIssuance
 }
+import io.iohk.atala.prism.management.console.repositories.metrics.CredentialIssuancesRepositoryMetrics
 import io.iohk.atala.prism.management.console.validations.CredentialDataValidator
 import org.slf4j.{Logger, LoggerFactory}
 import tofu.higherKind.Mid
@@ -334,30 +334,4 @@ private final class CredentialIssuancesRepositoryImpl[F[_]: BracketThrow](xa: Tr
       .logSQLErrors(s"getting credential, issuance id - $credentialIssuanceId", logger)
       .transact(xa)
   }
-}
-
-private final class CredentialIssuancesRepositoryMetrics[F[_]: TimeMeasureMetric: BracketThrow]
-    extends CredentialIssuancesRepository[Mid[F, *]] {
-
-  private val repoName = "CredentialIssuancesRepository"
-  private lazy val createTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "create")
-  private lazy val createBulkTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "createBulk")
-  private lazy val getTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "get")
-
-  override def create(
-      participantId: ParticipantId,
-      createCredentialIssuance: CreateCredentialIssuance
-  ): Mid[F, Either[ManagementConsoleError, CredentialIssuance.Id]] = _.measureOperationTime(createTimer)
-
-  override def createBulk(
-      participantId: ParticipantId,
-      credentialsType: CredentialTypeId,
-      issuanceName: String,
-      drafts: List[CreateCredentialBulk.Draft]
-  ): Mid[F, Either[ManagementConsoleError, CredentialIssuance.Id]] = _.measureOperationTime(createBulkTimer)
-
-  override def get(
-      credentialIssuanceId: CredentialIssuance.Id,
-      institutionId: ParticipantId
-  ): Mid[F, CredentialIssuance] = _.measureOperationTime(getTimer)
 }
