+++ prism-backend/management-console/src/main/scala/io/iohk/atala/prism/management/console/repositories/ParticipantsRepository.scala
@@ -20,8 +20,8 @@ import io.iohk.atala.prism.management.console.repositories.ParticipantsRepositor
   UpdateParticipantProfileRequest
 }
 import io.iohk.atala.prism.management.console.repositories.daos.{CredentialTypeDao, ParticipantsDAO}
-import io.iohk.atala.prism.metrics.{TimeMeasureMetric, TimeMeasureUtil}
-import io.iohk.atala.prism.metrics.TimeMeasureUtil.MeasureOps
+import io.iohk.atala.prism.management.console.repositories.metrics.ParticipantsRepositoryMetrics
+import io.iohk.atala.prism.metrics.TimeMeasureMetric
 import io.iohk.atala.prism.utils.syntax.DBConnectionOps
 import org.postgresql.util.PSQLException
 import org.slf4j.{Logger, LoggerFactory}
@@ -125,25 +125,4 @@ private final class ParticipantsRepositoryImpl[F[_]: BracketThrow](
     ParticipantsDAO
       .updateParticipantByID(request.id, request.participantProfile)
       .transact(xa)
-}
-
-private final class ParticipantsRepositoryMetrics[F[_]: TimeMeasureMetric: BracketThrow]
-    extends ParticipantsRepository[Mid[F, *]] {
-
-  private val repoName = "ParticipantsRepository"
-  private lazy val createTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "create")
-  private lazy val findByParticipantIdTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "findByParticipantId")
-  private lazy val findByDidTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "findByDid")
-  private lazy val updateTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "update")
-
-  override def create(request: CreateParticipantRequest): Mid[F, Either[ManagementConsoleError, Unit]] =
-    _.measureOperationTime(createTimer)
-
-  override def findBy(id: ParticipantId): Mid[F, Either[ManagementConsoleError, ParticipantInfo]] =
-    _.measureOperationTime(findByParticipantIdTimer)
-
-  override def findBy(did: DID): Mid[F, Either[ManagementConsoleError, ParticipantInfo]] =
-    _.measureOperationTime(findByDidTimer)
-
-  override def update(request: UpdateParticipantProfileRequest): Mid[F, Unit] = _.measureOperationTime(updateTimer)
 }
