+++ prism-backend/management-console/src/main/scala/io/iohk/atala/prism/management/console/repositories/ReceivedCredentialsRepository.scala
@@ -13,8 +13,8 @@ import io.iohk.atala.prism.management.console.models.{
 }
 import io.iohk.atala.prism.management.console.repositories.daos.ReceivedCredentialsDAO
 import io.iohk.atala.prism.management.console.repositories.daos.ReceivedCredentialsDAO.ReceivedSignedCredentialData
-import io.iohk.atala.prism.metrics.{TimeMeasureMetric, TimeMeasureUtil}
-import io.iohk.atala.prism.metrics.TimeMeasureUtil.MeasureOps
+import io.iohk.atala.prism.management.console.repositories.metrics.ReceivedCredentialsRepositoryMetrics
+import io.iohk.atala.prism.metrics.TimeMeasureMetric
 import io.iohk.atala.prism.utils.syntax.DBConnectionOps
 import org.slf4j.{Logger, LoggerFactory}
 import tofu.higherKind.Mid
@@ -67,26 +67,4 @@ private final class ReceivedCredentialsRepositoryImpl[F[_]: BracketThrow](xa: Tr
       .getLatestCredentialExternalId(verifierId)
       .logSQLErrors(s"getting latest credential external id, verifier id -  $verifierId", logger)
       .transact(xa)
-}
-
-private final class ReceivedCredentialsRepositoryMetrics[F[_]: TimeMeasureMetric: BracketThrow]
-    extends ReceivedCredentialsRepository[Mid[F, *]] {
-
-  private val repoName = "ReceivedCredentialsRepository"
-  private lazy val getCredentialsForTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "getCredentialsFor")
-  private lazy val createReceivedCredentialTimer =
-    TimeMeasureUtil.createDBQueryTimer(repoName, "createReceivedCredential")
-  private lazy val getLatestCredentialExternalIdTimer =
-    TimeMeasureUtil.createDBQueryTimer(repoName, "getLatestCredentialExternalId")
-
-  override def getCredentialsFor(
-      verifierId: ParticipantId,
-      contactId: Option[Contact.Id]
-  ): Mid[F, List[ReceivedSignedCredential]] = _.measureOperationTime(getCredentialsForTimer)
-
-  override def createReceivedCredential(data: ReceivedSignedCredentialData): Mid[F, Unit] =
-    _.measureOperationTime(createReceivedCredentialTimer)
-
-  override def getLatestCredentialExternalId(verifierId: ParticipantId): Mid[F, Option[CredentialExternalId]] =
-    _.measureOperationTime(getLatestCredentialExternalIdTimer)
 }
