+++ prism-backend/management-console/src/main/scala/io/iohk/atala/prism/management/console/repositories/RequestNoncesRepository.scala
@@ -8,8 +8,8 @@ import doobie.util.transactor.Transactor
 import io.iohk.atala.prism.auth.model.RequestNonce
 import io.iohk.atala.prism.management.console.models.ParticipantId
 import io.iohk.atala.prism.management.console.repositories.daos.RequestNoncesDAO
-import io.iohk.atala.prism.metrics.{TimeMeasureMetric, TimeMeasureUtil}
-import io.iohk.atala.prism.metrics.TimeMeasureUtil.MeasureOps
+import io.iohk.atala.prism.management.console.repositories.metrics.RequestNoncesRepositoryMetrics
+import io.iohk.atala.prism.metrics.TimeMeasureMetric
 import io.iohk.atala.prism.utils.syntax.DBConnectionOps
 import org.slf4j.{Logger, LoggerFactory}
 import tofu.higherKind.Mid
@@ -38,12 +38,4 @@ private final class RequestNoncesRepositoryPostgresImpl[F[_]: BracketThrow](xa: 
       .burn(participantId, requestNonce)
       .logSQLErrors(s"burning, participant id - $participantId", logger)
       .transact(xa)
-}
-
-private final class RequestNoncesRepositoryMetrics[F[_]: TimeMeasureMetric: BracketThrow]
-    extends RequestNoncesRepository[Mid[F, *]] {
-  private val repoName = "RequestNoncesRepositoryPostgresImpl"
-  private lazy val burnTimer = TimeMeasureUtil.createDBQueryTimer(repoName, "burn")
-  override def burn(participantId: ParticipantId, requestNonce: RequestNonce): Mid[F, Unit] =
-    _.measureOperationTime(burnTimer)
 }
